feat: integrate Param struct into FunctionDecl and remove Param.hpp
